+++ ollama-holes-plugin.cabal
@@ -128,6 +128,7 @@ library
       GHC.Plugin.OllamaHoles.Candidate
       GHC.Plugin.OllamaHoles.Candidate.Compat
       GHC.Plugin.OllamaHoles.Candidate.Rewrite
+      GHC.Plugin.OllamaHoles.Constants
       GHC.Plugin.OllamaHoles.Data.Flags
       GHC.Plugin.OllamaHoles.Data.Flags.Error
       GHC.Plugin.OllamaHoles.Data.Flags.Parse
@@ -171,6 +172,8 @@ library
       GHC.Plugin.OllamaHoles.Flags
       GHC.Plugin.OllamaHoles.Logger
       GHC.Plugin.OllamaHoles.Prompt
+      GHC.Plugin.OllamaHoles.Runtime
+      GHC.Plugin.OllamaHoles.Runtime.Ops
 
 
 
@@ -187,6 +190,7 @@ library ollama-holes-lib
     , GHC.Plugin.OllamaHoles.Candidate
     , GHC.Plugin.OllamaHoles.Candidate.Compat
     , GHC.Plugin.OllamaHoles.Candidate.Rewrite
+    , GHC.Plugin.OllamaHoles.Constants
     , GHC.Plugin.OllamaHoles.Data.Flags
     , GHC.Plugin.OllamaHoles.Data.Flags.Error
     , GHC.Plugin.OllamaHoles.Data.Flags.Parse
@@ -230,6 +234,8 @@ library ollama-holes-lib
     , GHC.Plugin.OllamaHoles.Flags
     , GHC.Plugin.OllamaHoles.Logger
     , GHC.Plugin.OllamaHoles.Prompt
+    , GHC.Plugin.OllamaHoles.Runtime
+    , GHC.Plugin.OllamaHoles.Runtime.Ops
