+++ src/GHC/Plugin/OllamaHoles/Runtime.hs
@@ -1,0 +1,5 @@
+module GHC.Plugin.OllamaHoles.Runtime
+  ( module GHC.Plugin.OllamaHoles.Runtime.Ops
+  ) where
+
+import GHC.Plugin.OllamaHoles.Runtime.Ops
